[TRL-283] refactor: ScheduleQueryRepository 구조 개선
