+++ src/main/java/com/cosain/trilo/trip/application/schedule/query/service/ScheduleDetailSearchService.java
@@ -21,7 +21,7 @@ public class ScheduleDetailSearchService implements ScheduleDetailSearchUseCase 
     }
 
     private ScheduleDetail findSchedule(Long scheduleId) {
-        return scheduleQueryRepository.findScheduleDetailByScheduleId(scheduleId).orElseThrow(ScheduleNotFoundException::new);
+        return scheduleQueryRepository.findScheduleDetailById(scheduleId).orElseThrow(ScheduleNotFoundException::new);
     }
 
 }
