+++ src/test/java/com/cosain/trilo/support/RepositoryTest.java
@@ -2,7 +2,7 @@ package com.cosain.trilo.support;
 
 import com.cosain.trilo.config.QueryDslConfig;
 import com.cosain.trilo.trip.infra.repository.day.DayQueryRepository;
-import com.cosain.trilo.trip.infra.repository.schedule.ScheduleQueryRepositoryImpl;
+import com.cosain.trilo.trip.infra.repository.schedule.ScheduleQueryRepository;
 import com.cosain.trilo.trip.infra.repository.trip.TripQueryRepositoryImpl;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
@@ -19,6 +19,6 @@ import java.lang.annotation.Target;
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @DataJpaTest
 @DirtiesContext
-@Import({QueryDslConfig.class, TripQueryRepositoryImpl.class, ScheduleQueryRepositoryImpl.class, DayQueryRepository.class})
+@Import({QueryDslConfig.class, TripQueryRepositoryImpl.class, ScheduleQueryRepository.class, DayQueryRepository.class})
 public @interface RepositoryTest {
 }
